[SPARK-59310][SQL] Address review comments for GroupPartitionsExec grouping metrics
Narrow the replicated-reads metric registration to expected keys that actually
carry multiple slots and rename its key to numReplicatedPartitionReads, make
the metrics map the single source of truth for what sendDriverMetrics posts,
drop the redundant numInputPartitions field from PartitionGrouping in favor of
the child's outputPartitioning, move the end-to-end metric tests to
KeyGroupedPartitioningSuite and run them with AQE on and off, document the
node-specific metrics in the web UI docs, and make ExecutableKeyedLeaf top
level.
